consolidate blob storage tests
