+++ tests/Microsoft.Bot.Builder.Azure.Tests/AzureBlobStorageTests.cs
@@ -2,11 +2,7 @@
 // Licensed under the MIT License.
 
 using System;
-using System.Collections.Generic;
-using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.Bot.Builder.Tests;
-using Microsoft.Bot.Schema;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Microsoft.WindowsAzure.Storage;
 using Microsoft.WindowsAzure.Storage.Blob;
@@ -15,13 +11,11 @@ using Newtonsoft.Json;
 namespace Microsoft.Bot.Builder.Azure.Tests
 {
     [TestClass]
-    public class AzureBlobStorageTests : StorageBaseTests
+    public class AzureBlobStorageTests : BlobStorageBaseTests
     {
-        private const string ConnectionString = @"AccountName=devstoreaccount1;AccountKey=Eby8vdM02xNOcqFlqUwJPLlmEtlCDXJ1OUzFT50uSRZ6IFsuFq2UVErCz4I6tq/K1SZFPTOtr/KBHBeksoGMGw==;DefaultEndpointsProtocol=http;BlobEndpoint=http://127.0.0.1:10000/devstoreaccount1;QueueEndpoint=http://127.0.0.1:10001/devstoreaccount1;TableEndpoint=http://127.0.0.1:10002/devstoreaccount1;";
-
         public TestContext TestContext { get; set; }
 
-        public string ContainerName
+        protected override string ContainerName
         {
             get
             {
@@ -56,7 +50,7 @@ namespace Microsoft.Bot.Builder.Azure.Tests
         }
 
         [TestMethod]
-        public async Task BlobStorageParamTest()
+        public void BlobStorageParamTest()
         {
             if (StorageEmulatorHelper.CheckEmulator())
             {
@@ -78,172 +72,7 @@ namespace Microsoft.Bot.Builder.Azure.Tests
             }
         }
 
-        [TestMethod]
-        public async Task TestBlobStorageWriteRead()
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                // Arrange
-                var storage = GetStorage();
-
-                var changes = new Dictionary<string, object>
-                {
-                    { "x", "hello" },
-                    { "y", "world" },
-                };
-
-                // Act
-                await storage.WriteAsync(changes);
-                var result = await storage.ReadAsync(new[] { "x", "y" });
-
-                // Assert
-                Assert.AreEqual(2, result.Count);
-                Assert.AreEqual("hello", result["x"]);
-                Assert.AreEqual("world", result["y"]);
-            }
-        }
-
-        [TestMethod]
-        public async Task TestBlobStorageWriteDeleteRead()
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                // Arrange
-                var storage = GetStorage();
-
-                var changes = new Dictionary<string, object>
-                {
-                    { "x", "hello" },
-                    { "y", "world" },
-                };
-
-                // Act
-                await storage.WriteAsync(changes);
-                await storage.DeleteAsync(new[] { "x" });
-                var result = await storage.ReadAsync(new[] { "x", "y" });
-
-                // Assert
-                Assert.AreEqual(1, result.Count);
-                Assert.AreEqual("world", result["y"]);
-            }
-        }
-
-        [TestMethod]
-        public async Task TestBlobStorageChanges()
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                // Arrange
-                var storage = GetStorage();
-
-                // Act
-                await storage.WriteAsync(new Dictionary<string, object> { { "a", "1.0" }, { "b", "2.0" } });
-                await storage.WriteAsync(new Dictionary<string, object> { { "c", "3.0" } });
-                await storage.DeleteAsync(new[] { "b" });
-                await storage.WriteAsync(new Dictionary<string, object> { { "a", "1.1" } });
-                var result = await storage.ReadAsync(new[] { "a", "b", "c", "d", "e" });
-
-                // Assert
-                Assert.AreEqual(2, result.Count);
-                Assert.AreEqual("1.1", result["a"]);
-                Assert.AreEqual("3.0", result["c"]);
-            }
-        }
-
-        [TestMethod]
-        public async Task TestConversationStateBlobStorage()
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                // Arrange
-                var storage = GetStorage();
-                var conversationState = new ConversationState(storage);
-                var propAccessor = conversationState.CreateProperty<Prop>("prop");
-
-                var adapter = new TestStorageAdapter();
-                var activity = new Activity
-                {
-                    ChannelId = "123",
-                    Conversation = new ConversationAccount { Id = "abc" },
-                };
-
-                // Act
-                var turnContext1 = new TurnContext(adapter, activity);
-                var propValue1 = await propAccessor.GetAsync(turnContext1, () => new Prop());
-                propValue1.X = "hello";
-                propValue1.Y = "world";
-                await conversationState.SaveChangesAsync(turnContext1, force: true);
-
-                var turnContext2 = new TurnContext(adapter, activity);
-                var propValue2 = await propAccessor.GetAsync(turnContext2);
-
-                // Assert
-                Assert.AreEqual("hello", propValue2.X);
-                Assert.AreEqual("world", propValue2.Y);
-
-                await propAccessor.DeleteAsync(turnContext1);
-                await conversationState.SaveChangesAsync(turnContext1);
-            }
-        }
-
-        [TestMethod]
-        public async Task TestConversationStateBlobStorage_TypeNameHandlingDefault()
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                await TestConversationStateBlobStorage_Method(GetStorage());
-            }
-        }
-
-        [TestMethod]
-        public async Task TestConversationStateBlobStorage_TypeNameHandlingNone()
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                await TestConversationStateBlobStorage_Method(GetStorage(true));
-            }
-        }
-
-        [TestMethod]
-        public async Task StatePersistsThroughMultiTurn_TypeNameHandlingNone()
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                await StatePersistsThroughMultiTurn(GetStorage(true));
-            }
-        }
-
-        private async Task TestConversationStateBlobStorage_Method(AzureBlobStorage storage)
-        {
-            if (StorageEmulatorHelper.CheckEmulator())
-            {
-                // Arrange
-                var conversationState = new ConversationState(storage);
-                var propAccessor = conversationState.CreateProperty<Prop>("prop");
-                var adapter = new TestStorageAdapter();
-                var activity = new Activity
-                {
-                    ChannelId = "123",
-                    Conversation = new ConversationAccount { Id = "abc" },
-                };
-
-                // Act
-                var turnContext1 = new TurnContext(adapter, activity);
-                var propValue1 = await propAccessor.GetAsync(turnContext1, () => new Prop());
-                propValue1.X = "hello";
-                propValue1.Y = "world";
-                await conversationState.SaveChangesAsync(turnContext1, force: true);
-
-                var turnContext2 = new TurnContext(adapter, activity);
-                var propValue2 = await propAccessor.GetAsync(turnContext2);
-
-                // Assert
-                Assert.AreEqual("hello", propValue2.X);
-                Assert.AreEqual("world", propValue2.Y);
-            }
-        }
-
-        private AzureBlobStorage GetStorage(bool typeNameHandlingNone = false)
+        protected override IStorage GetStorage(bool typeNameHandlingNone = false)
         {
             var storageAccount = CloudStorageAccount.Parse(ConnectionString);
             if (typeNameHandlingNone)
@@ -257,33 +86,6 @@ namespace Microsoft.Bot.Builder.Azure.Tests
             {
                 return new AzureBlobStorage(storageAccount, ContainerName);
             }
-        }
-
-        private class TestStorageAdapter : BotAdapter
-        {
-            public override Task DeleteActivityAsync(ITurnContext turnContext, ConversationReference reference, CancellationToken cancellationToken)
-            {
-                throw new NotImplementedException();
-            }
-
-            public override Task<ResourceResponse[]> SendActivitiesAsync(ITurnContext turnContext, Activity[] activities, CancellationToken cancellationToken)
-            {
-                throw new NotImplementedException();
-            }
-
-            public override Task<ResourceResponse> UpdateActivityAsync(ITurnContext turnContext, Activity activity, CancellationToken cancellationToken)
-            {
-                throw new NotImplementedException();
-            }
-        }
-
-        private class Prop : IStoreItem
-        {
-            public string X { get; set; }
-
-            public string Y { get; set; }
-
-            string IStoreItem.ETag { get; set; }
         }
     }
 }
